+++ Projects/DX_ActionBattle/Data/Script/Scene/TutorialScene.cpp
@@ -859,7 +859,7 @@ void TutorialScene::TutorialDraw()
 void TutorialScene::CameraUpdate()
 {
 	// カメラ関係
-	int size = m_objects.size();
+	int size = static_cast<int>(m_objects.size());
 	m_camera.SetTargetPos(m_objects["Player"]->GetPos(), VZero());
 
 	if (m_nowStageNum[0] == 0)
